+++ composables/technology/utils.js
@@ -1,28 +1,30 @@
 import { useTechnologyStore } from '@/stores/TechnologyStore';
 
-export const onAdd = async (name) => {
+export const onAdd = async (technology) => {
   const technologyStore = useTechnologyStore();
 
   await technologyStore.addTechnology(
-    { name: name }
+    { name: technology.name }
   );
   
   await msgAndRedirect('Technology has been added');
 }
 
-export const onEdit = async (id, name) => {
+export const onEdit = async (id, technology) => {
   const technologyStore = useTechnologyStore();
   
   await technologyStore.editTechnology(
     id,
-    { name: name }
+    { name: technology.name }
   );
   await msgAndRedirect('Technology has been edited');
 }
 
-export const onDelete = async () => {
-  const technologyStore = useTechnologyStore();
-  await technologyStore.deleteTechnology(id);
-  
-  await msgAndRedirect('Technology has been deleted');
+export const onDelete = async (doDelete) => {
+  if (doDelete === true) {
+    const technologyStore = useTechnologyStore();
+    await technologyStore.deleteTechnology(id);
+    
+    await msgAndRedirect('Technology has been deleted');
+  }
 }
