+++ composables/type/utils.js
@@ -1,28 +1,30 @@
 import { useTypeStore } from '@/stores/TypeStore';
 
-export const onAdd = async (name) => {
+export const onAdd = async (type) => {
   const typeStore = useTypeStore();
 
   await typeStore.addType(
-    { name: name }
+    { name: type.name }
   );
   
   await msgAndRedirect('Type has been added');
 }
 
-export const onEdit = async (id, name) => {
+export const onEdit = async (id, type) => {
   const typeStore = useTypeStore();
   
   await typeStore.editType(
     id,
-    { name: name }
+    { name: type.name }
   );
   await msgAndRedirect('Type has been edited');
 }
 
-export const onDelete = async () => {
-  const typeStore = useTypeStore();
-  await typeStore.deleteType(id);
-  
-  await msgAndRedirect('Type has been deleted');
+export const onDelete = async (doDelete) => {
+  if (doDelete === true) {
+    const typeStore = useTypeStore();
+    await typeStore.deleteType(id);
+    
+    await msgAndRedirect('Type has been deleted');
+  }
 }
